refactor: create base document class
